new BlogElement base class
